🚚 Rename: 프로필 페이지 파일 구조 분리 (#227)
