+++ src/pages/ProfilePage/UserInfo.jsx
@@ -1,4 +1,5 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
+import styled from 'styled-components';
 import { useNavigate } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 import { accountAtom } from '../../atoms/UserAtom';
@@ -6,17 +7,7 @@ import FollowCount from './FollowCount';
 import basicProfile from '../../assets/images/basic-profile-l.svg';
 import ButtonContainer from '../../components/common/Button/ButtonContainer';
 import { postFollow, deleteFollow } from '../../utils/Apis';
-import {
-  Container,
-  Header,
-  ProfileImage,
-  Profile,
-  UserName,
-  AccountName,
-  Introduction,
-  Interaction,
-  SelectionBox,
-} from './style/UserInfoStyle';
+
 
 const UserInfo = ({ data, myProfile }) => {
   const navigate = useNavigate();
@@ -95,3 +86,67 @@ const UserInfo = ({ data, myProfile }) => {
 };
 
 export default UserInfo;
+
+ const Container = styled.section`
+  text-align: center;
+`;
+
+const Header = styled.div`
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+  margin-top: 1.4rem;
+  padding: 1.6rem;
+`;
+
+ const ProfileImage = styled.div`
+  img {
+    width: 11rem;
+    height: 11rem;
+    object-fit: cover;
+    border-radius: 50%;
+  }
+`;
+
+ const Profile = styled.ul`
+  text-align: center;
+`;
+
+const UserName = styled.li`
+  font-weight: bold;
+  font-size: ${({ theme }) => theme.fontSize.large};
+`;
+
+const AccountName = styled.li`
+  color: ${({ theme }) => theme.colors.textColor};
+  font-size: 1.2rem;
+  margin-top: 0.6rem;
+`;
+
+const Introduction = styled.li`
+  margin-top: 1.6rem;
+  font-size: ${({ theme }) => theme.fontSize.medium};
+  color: ${({ theme }) => theme.colors.textColor};
+`;
+
+const Interaction = styled.div`
+  display: flex;
+  justify-content: center;
+  margin-top: 2.4rem;
+  gap: 1rem;
+  margin-bottom: 2.6rem;
+`;
+
+const SelectionBox = styled.div`
+  background-color: #f6f5f6;
+  box-shadow: 0 0 0.15rem rgba(0, 0, 0, 0.2);
+  color: ${({ theme }) => theme.colors.textColor};
+  height: 3.4rem;
+  text-align: center;
+  font-weight: 400;
+  font-size: ${({ theme }) => theme.fontSize.xsmall};
+  padding-top: 1.1rem;
+  strong {
+    color: #6521d3;
+  }
+`;
