--- src/pages/ProfilePage/style/FollowCountStyle.jsx
@@ -1,16 +1,0 @@
-import styled from 'styled-components';
-
-export const BoldText = styled.div`
-  font-weight: bold;
-  font-size: ${({ theme }) => theme.fontSize.large};
-`;
-export const BoldBlurText = styled.div`
-  font-weight: bold;
-  color: ${({ theme }) => theme.colors.textColor};
-  font-size: ${({ theme }) => theme.fontSize.xlarge};
-`;
-export const FollowText = styled.div`
-  font-size: ${({ theme }) => theme.fontSize.medium};
-  color: ${({ theme }) => theme.colors.textColor};
-  margin-top: 0.6rem;
-`;
